Unit tests improvements.
These code changes improve the unit tests by the re-use of the
DisposableApplication class.
