+++ OfficeToPDF.Tests/OfficeToPDF.Tests.csproj
@@ -137,6 +137,7 @@
     <Compile Include="ArgParserTests.cs" />
     <Compile Include="COMServerTests.cs" />
     <Compile Include="ConverterFactoryTests.cs" />
+    <Compile Include="DisposableApplication.cs" />
     <Compile Include="Functional.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="RandomNumberGenerator.cs" />
